% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkBasic}
\alias{checkBasic}
\title{Checks Inputs}
\usage{
checkBasic(s, C, duration, times)
}
\arguments{
\item{s}{activity center of a single individual}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{times}{times at which to evaluate hazards and their integrals}
}
\description{
Checks if inputs for a hazard or simulate function are correct
}
