% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{checkCaptHist}
\alias{checkCaptHist}
\title{Check Capture History}
\usage{
checkCaptHist(capthist, duration, K)
}
\arguments{
\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}

\item{duration}{duration of survey}

\item{K}{number of cameras or traps in a survey}
}
\description{
Checks if a supplied capthist matrix is in the expected format
}
