% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{convertCaptHist}
\alias{convertCaptHist}
\title{Converts individual ids}
\usage{
convertCaptHist(ids, N)
}
\arguments{
\item{ids}{id number of the individuals captured at detections}

\item{N}{is total population size}
}
\value{
ids, giving the converted individual IDs, and new_order, how the individuals should be ordered for use with timeRescaling() and runPPC()
}
\description{
Converts capthist from those produced by simulation functions
Converts capthist to the form required by CheckCaptHist()
In particular, individuals must be numbered 1:m without unused ids
}
\details{
The simulate_* functions used the row numbers of the activity center matrix S to number ids.
This results in individuals with integers in the range of 1:N, but when there are undetected individuals, there are unused numbers.
The timeRescaling() and runPPC() functions require detected individuals to have the lowest numbers, starting at one.
}
