% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{empiricalAC}
\alias{empiricalAC}
\title{Get a starting estimate for activity centers}
\usage{
empiricalAC(C, capthist)
}
\arguments{
\item{C}{matrix of camera or trap locations}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
A weighted average of the locations where an individual was observed
}
