% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist.R
\name{findAC}
\alias{findAC}
\title{Find the activity center for a single individual}
\usage{
findAC(s, C, duration, pars, hazard, capthist)
}
\arguments{
\item{s}{starting point for the activity center}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
Applies Bayes' Rule to return a point estimate for an observed individual
}
\details{
Note there are no constraints to ensure that the estimated ACs are within the survey region
}
