% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist.R
\name{findACs}
\alias{findACs}
\title{Find the activity center for all individuals}
\usage{
findACs(S, C, duration, pars, hazard, N, mask, capthist)
}
\arguments{
\item{S}{starting points for the activity centers of observed individuals}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{N}{estimated population size}

\item{mask}{(only needed if \code{S} not provided) points that lie within survey region and which serve as an approximation of it}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
For observed individuals, use maximum likelihood estimation
For unobserved individuals, sample from the likelihood.
}
