% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{FT_Stat}
\alias{FT_Stat}
\title{Freeman-Tukey statistic}
\usage{
FT_Stat(obs, exp)
}
\arguments{
\item{obs}{The observed number of events, either a vector of matrix}

\item{exp}{The expected number of events, with the same dimensions}
}
\description{
Returns the Freeman-Tukey statistic, without the coefficient of 4.
}
