% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{FTMetrics}
\alias{FTMetrics}
\title{FT metrics}
\usage{
FTMetrics(data, capthist, N, K, duration = NULL)
}
\arguments{
\item{data}{pre-calculated information needed to compute metric}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}

\item{N}{estimated population size}

\item{K}{number of cameras or traps in a survey}

\item{duration}{duration of survey}
}
\description{
FT metrics discussed in Choo et al, 2024 and other works.
Can serve as the \code{metric} input to \code{runPPC()}.
}
\details{
\code{duration} is not required for this metric
}
