% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist.R
\name{getResiduals}
\alias{getResiduals}
\title{Residuals for Frequentist SCR Models}
\usage{
getResiduals(
  S = NULL,
  C,
  duration,
  pars,
  hazard,
  N,
  comps = NULL,
  mask,
  capthist
)
}
\arguments{
\item{S}{matrix of activity centers. If not provided, \code{findACs()} will be used to estimate them}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{N}{estimated population size}

\item{comps}{(optional) output of \code{Hazards()} function.}

\item{mask}{(only needed if \code{S} not provided) points that lie within survey region and which serve as an approximation of it}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\value{
a list with three elements: \cr
\itemize{
\item \code{S} a Nx2 matrix of activity centers used for the tests \cr
\item \code{resid_trap} the trap residuals \cr
\item \code{resid_ind} the individual residuals \cr
}
}
\description{
Calculates both trap and individual level residuals for frequentist SCR models
}
