% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazards.R
\name{hazard_cosine}
\alias{hazard_cosine}
\title{Cosine SCR Hazard function}
\usage{
hazard_cosine(s, C, duration, pars, times, capthist = NULL)
}
\arguments{
\item{s}{activity center of a single individual}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of pars: \eqn{\lambda_0}, \eqn{\sigma} and \eqn{\phi}}

\item{times}{times at which to evaluate hazards and their integrals}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\value{
two matrices with hazards and integrals, see details
}
\description{
Calculates the hazard and integrals of the hazard for a cosine wave detection function. See the PhD thesis of G. Distiller, 2016
}
\details{
Returns a list with elements \code{h} and \code{H}, each being a matrix with a row for each value of times
and a column for each trap. The values of \code{h} are hazards and \code{H} are integrals of \code{h} since time zero.

\eqn{\lambda_0} is the rate of detections at the activity center, and \eqn{\sigma} the home range size.
\eqn{\phi} is the magnitude of the wave, 0 = standard SCR and 1=max possible.
}
