% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hazards.R
\name{hazard_SESCR}
\alias{hazard_SESCR}
\title{SESCR Hazard function}
\usage{
hazard_SESCR(s, C, duration, pars, times, capthist)
}
\arguments{
\item{s}{activity center of a single individual}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of pars: vector of pars: \eqn{\lambda_0}, \eqn{\sigma}, \eqn{\beta} and \eqn{r}, see details}

\item{times}{times at which to evaluate hazards and their integrals}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
Hazard for the SESCR model of van Helsdingen & Jones-Todd, 2026
}
\details{
Returns a list with components \code{h} and \code{H}, each being a matrix with a row for each value of times
and a column for each trap. The values of \code{h} are hazards and \code{H} are integrals of \code{h} since time zero.

\eqn{\lambda_0} is the rate of detections at the activity center when there has been no prior detections. \eqn{\sigma} is the home range size.
\eqn{\beta} is the temporal decay rate for self-excitement and \eqn{r} the spatial decay.
}
