% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Hazards}
\alias{Hazards}
\title{Calculate hazard integrals}
\usage{
Hazards(S, C, duration, pars, times, hazard, N, capthist)
}
\arguments{
\item{S}{matrix of activity centers. If not provided, \code{findACs()} will be used to estimate them}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{times}{times at which to evaluate the hazard and its integral}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{N}{estimated population size}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
Output is the comps input of \code{timeRescaling()} and \code{residuals()} functions. For multiple individuals.
}
