% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{Integrals}
\alias{Integrals}
\title{Calculate hazard integrals at end of survey,}
\usage{
Integrals(S, C, duration, pars, hazard, N, capthist)
}
\arguments{
\item{S}{matrix of activity centers. If not provided, \code{findACs()} will be used to estimate them}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{N}{estimated population size}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
Returns hazard integrals for each pair of individual and camera
}
