% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{residualPlot}
\alias{residualPlot}
\title{Make a visually meaningful residual plot}
\usage{
residualPlot(resids)
}
\arguments{
\item{resids}{the residuals}
}
\description{
Make a visually meaningful residual plot
}
