% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{runPPC}
\alias{runPPC}
\title{Run Bayesian PPCs for SCR Models}
\usage{
runPPC(
  S,
  C,
  duration,
  samples,
  Z,
  hazard,
  simulate,
  metric,
  prep,
  mode = 0,
  n_metrics = 1,
  capthist
)
}
\arguments{
\item{S}{matrix of activity center estimates. X}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{samples}{matrix of parameter estimates}

\item{Z}{matrix of data augmentation binary variables}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{simulate}{function that simulates model and returns a \code{capthist} object}

\item{metric}{function that returns one or more scalar test statistics}

\item{prep}{auxillary function needed for modes 1 and 2}

\item{mode}{Type of PPC metric, see details}

\item{n_metrics}{number of scalar metrics to compute}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
Runs a Bayesian PPC after MCMC has been run and a \code{metric} indicated
}
\details{
Regardless of \code{mode}, the metric is computed on both simulated and the actual \code{capthist} \cr
When \code{mode=0}, the metric depends only on a \code{capthist} object. \cr
When \code{mode=1} or \code{mode=2}, the metric depends on \code{capthist} and the output of \code{prep} \cr
When \code{mode=2}, \code{prep} requires a further simulated \code{capthist}
}
