% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{SESCROgata}
\alias{SESCROgata}
\title{Helper function for simulate_SESCR()}
\usage{
SESCROgata(mu, beta, d, mu0, K, t)
}
\arguments{
\item{mu}{baseline rate for each stream}

\item{beta}{temporal decay of self-excitement}

\item{d}{self-excitement over space (NxN)}

\item{mu0}{volume of baseline rate across all streams}

\item{K}{number of cameras and streams}

\item{t}{duration of survey}
}
\description{
Helper function that simulates from SESCR for a single individual
Not designed to be called except within simSESCR()
}
