% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulate_cosine}
\alias{simulate_cosine}
\title{Simulates continuous-time SCR with a cosine wave function}
\usage{
simulate_cosine(S, C, duration, pars)
}
\arguments{
\item{S}{matrix of activity centers}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of pars: \eqn{\lambda_0}, \eqn{\sigma} and \eqn{\phi}}
}
\value{
individuals are numbered according to their row in \code{S}.
}
\description{
All individuals and terrain homogeneous
}
