% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulate_SCR}
\alias{simulate_SCR}
\title{Simulates standard continuous-time SCR}
\usage{
simulate_SCR(S, C, duration, pars)
}
\arguments{
\item{S}{matrix of activity centers}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of pars: \eqn{\lambda_0} and \eqn{\sigma}, see details}
}
\value{
individuals are numbered according to their row in \code{S}.
}
\description{
All individuals and terrain homogeneous
}
