% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulate_SESCR}
\alias{simulate_SESCR}
\title{Simulates from the SESCR model}
\usage{
simulate_SESCR(S, C, duration, pars)
}
\arguments{
\item{S}{matrix of activity centers}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of pars: vector of pars: \eqn{\lambda_0}, \eqn{\sigma}, \eqn{\beta} and \eqn{r}, see details}
}
\value{
individuals are numbered according to their row in \code{S}.
}
\description{
Simulates from the SESCR model
}
