% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist.R
\name{singleNLL}
\alias{singleNLL}
\title{Negative log-likelihood of a continuous-time SCR model}
\usage{
singleNLL(s, C, duration, pars, hazard, capthist)
}
\arguments{
\item{s}{activity center of a single individual}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
Applies the generic likelihood formula for any Poisson SCR model where
individuals are independent
}
