% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{subSampleTest}
\alias{subSampleTest}
\title{Subsampling}
\usage{
subSampleTest(values, n = NULL)
}
\arguments{
\item{values}{sequence of compensator differences}

\item{n}{size of subsample}
}
\value{
p-value from Kolmogorov-Smirnov test on subsample, null is Exp(1)
}
\description{
Subsamples a sequence of numbers that under a null hypothesis are exponentially distributed
}
\details{
Size of subsample defaults to n^(2/3) as suggested by Reynaud-Bouret et al (2014)
}
