% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{trans_matrix}
\alias{trans_matrix}
\title{Transition Matrix for SESCR}
\usage{
trans_matrix(s, C, lambda0, beta, Dratio, sigma)
}
\arguments{
\item{s}{activity center of a single individual}

\item{C}{matrix of camera or trap locations}

\item{lambda0}{detection rate at activity center}

\item{beta}{temporal decay of self-excitement}

\item{Dratio}{spatial decay of self-excitement}

\item{sigma}{home range size}
}
\description{
Returns a matrix showing the expected number of "transitions" between cameras, for a single individual
Only valid for the SESCR Model. Called by \code{transMetric_prep_SESCR}.
}
