% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{transMetric_prep_SESCR}
\alias{transMetric_prep_SESCR}
\title{FT Transition Metric Preparation}
\usage{
transMetric_prep_SESCR(S, C, duration = NULL, pars, hazard, capthist)
}
\arguments{
\item{S}{matrix of activity centers. If not provided, \code{findACs()} will be used to estimate them}

\item{C}{matrix of camera or trap locations}

\item{duration}{duration of survey}

\item{pars}{vector of parameter values for \code{hazard}}

\item{hazard}{function that returns hazard rates, e.g. \code{hazard_SCR()}}

\item{capthist}{3-column matrix giving times, ids and trap ids for each detection.
\code{capthist} must be have continuous detection times in the first column,
individual IDs numbered 1:m in the second, and trap IDs numbered 1:K in the third}
}
\description{
For SESCR only. Transition matrix metric discussed in Appendix A
Can serve as the \code{prep} input to \code{runPPC()}.
}
